move `det_installed_python_packages`, `det_pip_version`, `run_pip_check` to Python easyblock, so they can also be used there
